minimum type-safety : object approach
