+++ src/services/cluster/actors/peer.rs
@@ -69,24 +69,3 @@ impl From<(OwnedWriteHalf, PeerKind)> for WriteConnected {
         WriteConnected { stream: w, kind: peer_kind }
     }
 }
-
-#[test]
-fn test_peer_state_serialization() {
-    use crate::services::query_io::QueryIO;
-
-    //GIVEN
-    let peer_state = PeerState { term: 1, offset: 2, last_updated: 3 };
-    //WHEN
-    let peer_state_serialized: QueryIO = peer_state.into();
-    let peer_state_serialized = peer_state_serialized.serialize();
-    //THEN
-    assert_eq!("^\r\n$1\r\n1\r\n$1\r\n2\r\n$1\r\n3\r\n", peer_state_serialized);
-
-    //GIVEN
-    let peer_state = PeerState { term: 5, offset: 3232, last_updated: 35535300 };
-    //WHEN
-    let peer_state_serialized: QueryIO = peer_state.into();
-    let peer_state_serialized = peer_state_serialized.serialize();
-    //THEN
-    assert_eq!("^\r\n$1\r\n5\r\n$4\r\n3232\r\n$8\r\n35535300\r\n", peer_state_serialized);
-}
